+++ tests/cli/test_agent.py
@@ -26,10 +26,10 @@ def test_agent_help():
 
 def test_agent_start_fails_no_token(monkeypatch):
     start = MagicMock()
-    monkeypatch.setattr("prefect.agent.local.LocalAgent.start", start)
+    monkeypatch.setattr("prefect.agent.docker.DockerAgent.start", start)
 
     docker_client = MagicMock()
-    monkeypatch.setattr("prefect.agent.local.agent.docker.APIClient", docker_client)
+    monkeypatch.setattr("prefect.agent.docker.agent.docker.APIClient", docker_client)
 
     runner = CliRunner()
     result = runner.invoke(agent, ["start"])
@@ -38,25 +38,25 @@ def test_agent_start_fails_no_token(monkeypatch):
 
 def test_agent_start_token(monkeypatch, runner_token):
     start = MagicMock()
-    monkeypatch.setattr("prefect.agent.local.LocalAgent.start", start)
+    monkeypatch.setattr("prefect.agent.docker.DockerAgent.start", start)
 
     docker_client = MagicMock()
-    monkeypatch.setattr("prefect.agent.local.agent.docker.APIClient", docker_client)
+    monkeypatch.setattr("prefect.agent.docker.agent.docker.APIClient", docker_client)
 
     runner = CliRunner()
-    result = runner.invoke(agent, ["start", "-t", "test"])
+    result = runner.invoke(agent, ["start", "docker", "-t", "test"])
     assert result.exit_code == 0
 
 
 def test_agent_start_verbose(monkeypatch, runner_token):
     start = MagicMock()
-    monkeypatch.setattr("prefect.agent.local.LocalAgent.start", start)
+    monkeypatch.setattr("prefect.agent.docker.DockerAgent.start", start)
 
     docker_client = MagicMock()
-    monkeypatch.setattr("prefect.agent.local.agent.docker.APIClient", docker_client)
+    monkeypatch.setattr("prefect.agent.docker.agent.docker.APIClient", docker_client)
 
     runner = CliRunner()
-    result = runner.invoke(agent, ["start", "-v"])
+    result = runner.invoke(agent, ["start", "docker", "-v"])
     assert result.exit_code == 0
 
 
@@ -64,11 +64,29 @@ def test_agent_start_local(monkeypatch, runner_token):
     start = MagicMock()
     monkeypatch.setattr("prefect.agent.local.LocalAgent.start", start)
 
-    docker_client = MagicMock()
-    monkeypatch.setattr("prefect.agent.local.agent.docker.APIClient", docker_client)
-
     runner = CliRunner()
     result = runner.invoke(agent, ["start", "local"])
+    assert result.exit_code == 0
+
+
+def test_agent_start_local_import_paths(monkeypatch, runner_token):
+    start = MagicMock()
+    monkeypatch.setattr("prefect.agent.local.LocalAgent.start", start)
+
+    runner = CliRunner()
+    result = runner.invoke(agent, ["start", "local", "-p", "test"])
+    assert result.exit_code == 0
+
+
+def test_agent_start_docker(monkeypatch, runner_token):
+    start = MagicMock()
+    monkeypatch.setattr("prefect.agent.docker.DockerAgent.start", start)
+
+    docker_client = MagicMock()
+    monkeypatch.setattr("prefect.agent.docker.agent.docker.APIClient", docker_client)
+
+    runner = CliRunner()
+    result = runner.invoke(agent, ["start", "docker"])
     assert result.exit_code == 0
 
 
@@ -144,50 +162,50 @@ def test_agent_start_fargate_kwargs_received(monkeypatch, runner_token):
 
 def test_agent_start_name(monkeypatch, runner_token):
     start = MagicMock()
-    monkeypatch.setattr("prefect.agent.local.LocalAgent.start", start)
+    monkeypatch.setattr("prefect.agent.docker.DockerAgent.start", start)
 
     docker_client = MagicMock()
-    monkeypatch.setattr("prefect.agent.local.agent.docker.APIClient", docker_client)
+    monkeypatch.setattr("prefect.agent.docker.agent.docker.APIClient", docker_client)
 
     runner = CliRunner()
-    result = runner.invoke(agent, ["start", "--name", "test_agent"])
+    result = runner.invoke(agent, ["start", "docker", "--name", "test_agent"])
     assert result.exit_code == 0
 
 
 def test_agent_start_local_vars(monkeypatch, runner_token):
     start = MagicMock()
-    monkeypatch.setattr("prefect.agent.local.LocalAgent.start", start)
+    monkeypatch.setattr("prefect.agent.docker.DockerAgent.start", start)
 
     docker_client = MagicMock()
-    monkeypatch.setattr("prefect.agent.local.agent.docker.APIClient", docker_client)
+    monkeypatch.setattr("prefect.agent.docker.agent.docker.APIClient", docker_client)
 
     runner = CliRunner()
     result = runner.invoke(
-        agent, ["start", "-t", "test", "--base-url", "url", "--no-pull"]
+        agent, ["start", "docker", "-t", "test", "--base-url", "url", "--no-pull"]
     )
     assert result.exit_code == 0
 
 
 def test_agent_start_local_labels(monkeypatch, runner_token):
     start = MagicMock()
-    monkeypatch.setattr("prefect.agent.local.LocalAgent.start", start)
+    monkeypatch.setattr("prefect.agent.docker.DockerAgent.start", start)
 
     docker_client = MagicMock()
-    monkeypatch.setattr("prefect.agent.local.agent.docker.APIClient", docker_client)
+    monkeypatch.setattr("prefect.agent.docker.agent.docker.APIClient", docker_client)
 
     runner = CliRunner()
     result = runner.invoke(
-        agent, ["start", "-t", "test", "--label", "label1", "-l", "label2"]
+        agent, ["start", "docker", "-t", "test", "--label", "label1", "-l", "label2"]
     )
     assert result.exit_code == 0
 
 
-def test_agent_start_fails(monkeypatch):
+def test_agent_start_fails(monkeypatch, runner_token):
     start = MagicMock()
-    monkeypatch.setattr("prefect.agent.local.LocalAgent.start", start)
+    monkeypatch.setattr("prefect.agent.docker.DockerAgent.start", start)
 
     docker_client = MagicMock()
-    monkeypatch.setattr("prefect.agent.local.agent.docker.APIClient", docker_client)
+    monkeypatch.setattr("prefect.agent.docker.agent.docker.APIClient", docker_client)
 
     runner = CliRunner()
     result = runner.invoke(agent, ["start", "TEST"])
@@ -195,11 +213,11 @@ def test_agent_start_fails(monkeypatch):
     assert "TEST is not a valid agent" in result.output
 
 
-def test_agent_install():
+def test_agent_install_local():
     runner = CliRunner()
-    result = runner.invoke(agent, ["install"])
+    result = runner.invoke(agent, ["install", "local"])
     assert result.exit_code == 0
-    assert "apiVersion" in result.output
+    assert "supervisord" in result.output
 
 
 def test_agent_install_kubernetes():
@@ -216,12 +234,13 @@ def test_agent_install_fails_non_valid_agent():
     assert "fake_agent is not a supported agent for `install`" in result.output
 
 
-def test_agent_install_passes_args():
+def test_agent_install_k8s_asses_args():
     runner = CliRunner()
     result = runner.invoke(
         agent,
         [
             "install",
+            "kubernetes",
             "--token",
             "TEST_TOKEN",
             "--api",
@@ -247,12 +266,13 @@ def test_agent_install_passes_args():
     assert "test_label2" in result.output
 
 
-def test_agent_install_no_resource_manager():
+def test_agent_install_k8s_no_resource_manager():
     runner = CliRunner()
     result = runner.invoke(
         agent,
         [
             "install",
+            "kubernetes",
             "--token",
             "TEST_TOKEN",
             "--api",
@@ -269,3 +289,27 @@ def test_agent_install_no_resource_manager():
     assert "TEST_NAMESPACE" in result.output
     assert not "resource-manager" in result.output
     assert "secret-test" in result.output
+
+
+def test_agent_install_local_asses_args():
+    runner = CliRunner()
+    result = runner.invoke(
+        agent,
+        [
+            "install",
+            "local",
+            "--token",
+            "TEST_TOKEN",
+            "--label",
+            "test_label1",
+            "-l",
+            "test_label2",
+            "--import-path",
+            "my_path",
+        ],
+    )
+    assert result.exit_code == 0
+    assert "TEST_TOKEN" in result.output
+    assert "test_label1" in result.output
+    assert "test_label2" in result.output
+    assert "my_path" in result.output