+++ tests/environments/storage/test_local_storage.py
@@ -13,13 +13,20 @@ from prefect.utilities.configuration import set_temporary_config
 def test_create_local_storage():
     storage = Local()
     assert storage
-    assert storage.directory.endswith(".prefect/flows")
+    end_path = os.path.join(".prefect", "flows")
+    assert storage.directory.endswith(end_path)
 
 
 def test_create_local_storage_with_custom_dir():
     storage = Local(directory=".")
     assert storage
     assert os.path.isabs(storage.directory)
+
+
+def test_create_local_storage_without_validation():
+    storage = Local(directory="C:\\Users\\chris\\.prefect\\flows", validate=False)
+    assert storage
+    assert storage.directory == "C:\\Users\\chris\\.prefect\\flows"
 
 
 def test_add_flow_to_storage():